requirements.txt
cbor==1.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
datasets==1.11.0
dill==0.3.4
filelock==3.0.12
fsspec==2021.8.1
huggingface-hub==0.0.16
idna==3.2
importlib-metadata==4.8.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
multiprocess==0.70.12.2
numpy==1.21.2
packaging==21.0
pandas==1.3.2
Pillow==8.3.1
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==5.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytrec-eval==0.5
pytz==2021.1
requests==2.26.0
six==1.16.0
torch==1.9.0
tqdm==4.62.2
trec-car-tools==2.5.4
typing-extensions==3.10.0.2
urllib3==1.26.6
xxhash==2.0.2
zipp==3.5.0
